Async alert loop; Filter by age
This commit does two things:
- Moves alerting into a goroutine that fires on an interval. Currently
that interval is hardcoded to 1m; ultimately it should be made
configurable via HiveHealthConfig.
- Adds filtering by the age of the sync failure, currently hardcoded to
4h; ultimately this should also be made configurable via
HiveHealthConfig.

I tracked the duplicated/garbled messaging in the previous commit down
to what can only be a bug in the logger code. I didn't get further than
that; but I refactored to work around it (putting all the text in the
message body rather than using the key/values). This will be n/a when
this code is replaced with real alerting anyway.